ref: move full state of panel out of panel component, create separate panel component
