+++ src/features/events_list/components/EventsPanel/EventsPanel.module.css
