+++ src/features/events_list/components/ShortState/ShortState.module.css
