+++ src/features/events_list/index.tsx
@@ -2,13 +2,11 @@ import { useAction, useAtom } from '@reatom/react';
 import { useCallback } from 'react';
 import { currentEventAtom } from '~core/shared_state';
 import { scheduledAutoFocus } from '~core/shared_state/currentEvent';
-import { eventListResourceAtom } from './atoms/eventListResource';
 import { autoSelectEvent } from './atoms/autoSelectEvent';
-import { EventsListPanel } from './components';
+import { EventsPanel } from './components/EventsPanel/EventsPanel';
 
 export function EventList() {
   const [currentEvent, currentEventActions] = useAtom(currentEventAtom);
-  const [eventListResource] = useAtom(eventListResourceAtom);
   const scheduleAutoFocus = useAction(scheduledAutoFocus.setTrue);
   const onCurrentChange = useCallback((id: string) => {
     currentEventActions.setCurrentEventId(id);
@@ -17,12 +15,6 @@ export function EventList() {
 
   useAtom(autoSelectEvent);
   return (
-    <EventsListPanel
-      current={currentEvent?.id ?? null}
-      error={eventListResource.error}
-      loading={eventListResource.loading}
-      eventsList={eventListResource.data}
-      onCurrentChange={onCurrentChange}
-    />
+    <EventsPanel currentEventId={currentEvent?.id} onCurrentChange={onCurrentChange} />
   );
 }
